edit ThemeToggle context and hook useTheme
