+++ app/src/components/ThemeToggle/ThemeToggle.tsx
@@ -1,12 +1,15 @@
-import useTheme from "@hooks/useTheme";
+// import useTheme from "@hooks/useTheme";
 import SunIcon from "@assets/icons/sun.svg?react";
 import MoonIcon from "@assets/icons/moon.svg?react";
 import "./ThemeToggle.scss";
+import { useContext } from "react";
+import ThemeToggleContext from "@context/ThemeToggle/ThemeToggle";
 
 const ThemeToggle = () => {
-  const [toggleTheme, theme] = useTheme();
+  // const [toggleTheme, theme] = useTheme();
+  const themeToggleCtx = useContext(ThemeToggleContext);
 
-  const isLightThemeActive = theme === "light";
+  const isLightThemeActive = themeToggleCtx.theme === "light";
 
   return (
     <div
@@ -14,7 +17,10 @@ const ThemeToggle = () => {
       data-theme={isLightThemeActive ? "light" : "dark"}
     >
       <SunIcon className="theme-icon" />
-      <button className="toggle-theme-button" onClick={toggleTheme}>
+      <button
+        className="toggle-theme-button"
+        onClick={themeToggleCtx.toggleTheme}
+      >
         <div className="toggle-theme"></div>
         <div
           className={`${
